Made 'renderers' pluggable in the graphite reporter
Because many metrics are somewhat verbose and because of the nature of
Graphite, it´s desirable to be able to decide which fields that are
sent to graphite.
E.g for a metered metric, perhaps the only really interesting field to
graph is the one minute rate. With a polling time of one minute, this
would generate a neat and relevant graph in graphite.