Removed LoggerMemoryLeakFix.
As painful as that was to find, it's not significant enough to dedicate a lot of
resources to (and it's also a JDK-level bug). I'd rather not accidentally introduce
a JDK implementation- or version-specific bug by trying to unfuck java.util.*.

We can always revive this in the future should it become a problem.