add support to override group and type for meter annotations
Metered, Timed, ExceptionMetered, and Guage have all been updated to support
two new attributes on the annotation:
- group - if omitted, defaults to the package name (as before)

- name - if omitted, defaults to simple class name (as before)

This involved changing the Guice annotation support code to check for these
attributes as well as the jersey dispatcher code for the same reason.
In order to make the behavior consistent across the two, the methods
for choosing the appropriate value have been placed into MetricName
as static methods.