Use the full MetricName when creating CSV files.
This reduces naming conflicts in the presence of "related" metrics.
Specifically, we add instrumentation to Guava's caches and end up with
metrics with names like <cacheName>.<hitRate|missRate> etc. Using either
just the name or just the scope results in errors while creating the CSV
files, if multiple instrumented caches are present.

Refs #212, #223.