feature: allow GraphiteReporter to not send all its extra metrics
Each intermediate metrics are currently sent automatically to Graphite
while it's an intrinsic Graphite feature. This causes many useless files
to be created on the Graphite server. This can be really annoying when
the number of metrics becomes huge.

This feature does not change the default behaviour of the module
but allow a fine-grained control over these metrics sending to the
Graphite server :

```
builder...
       .withoutMeanRates()
       .withoutPercentiles()
       .withoutStandardDeviation()
       .withoutMinMax()
       .withoutMean()
       .build();
```

A reset method is provided :

```
builder.withAllExtraReports();
```

And all extra reports can be deactivated in one buider method :

```
builder.withoutExtraReports();
```