Add support for patterns in MBean object names
If the object name is a pattern, perform a query for matching object
names and return the value of the given attribute if the pattern yields
a unique match. If the match is ambiguous, return null to avoid
returning the value from a wrong match.

The object name of an MBean consists of a domain part and a list of
key/value pairs that together form a unique identifier. The values,
however, often depend on current configuration parameters (i.e. the
hostname). By using an object name pattern it is possible to avoid the
need of considering configuration parameters when passing the object
name to the JmxAttributeGauge.

(cherry picked from commit 509d120)