httpasyncclient: Fix instrumentation of the InstrumentedNHttpClientBuilder
Previously the Timer only measured the creation of the Future but not
the execution of the request.
In order to time the actual request execution a custom `FutureCallback` is used.

Add a separate TimerTest that ensures that the Timer is stopped
correctly (i.e. after the response is received and not after the Future
is created).

This change might have a significant impact on current consumers as
the scope of what is measured changes.