Allow set custom prefix Jetty QueuedThreadPool's metrics
Prior to this fix all Jetty QueuedThreadPool's metrics are prefixed
by the class. The fix allows set the metric's name prefix. The origin value is
used as the default value because of backward compatibility. We don't want to
arbitrary chnage the name of metrics once existing users swith to new version of
the library.