Add basic sanitization for default CsvFileProvider
The current sanitization offered by the default CsvFileProvider does nothing.
Although this may be acceptable in most cases there are some known bad filename chars that we can restrict that cover
99% of the dropwizard user base (windows, unix, linux).

The forward slash '/' is illegal in all three OS and should be at a minimum sanitizied. The forward slash is a common
character in metric names if people are constructing metric names from URIs.