Negative and large tick value fixes
Collision buffer will cause overflow at 417 days of server uptime (not
JVM). Also it would take 1,000,000 requests in the same millisecond to
cause loss of accuracy of 1 ms, so this buffer is not needed and given
that it causes problems, should be removed.
Also added test case for large negative values of tick (rarely happens,
but according to nanoTime api, it can)