Add per-measurement level filtering of metrics
This configuration has existed previously, and still exists in metrics-reporter-config.
This commit provides support for it again within the GraphiteReporter.

This commit breaks compatibility as where MetricsFilter is used as an SPI by third-parties it now requires an additional method to be implemented.
This breakage does not exist in master (v4.0.0) as Java8 and a default interface method used.

ref: https://github.com/addthis/metrics-reporter-config/pull/2