Move closing` DatagramChannel` to the close method (#1080)
We can align `GraphiteUDP` with the TCP based `Graphite` and close the
socket in the `close` method instead of the connect method. The close
method will be invoked every time after a portion of metrics has been
sent, so we can close the socket there. We also do not need to call
the method `isConnected` during sending metrics because the socket is
opened right before sending data, so it can't be closed by timeout.

(cherry picked from commit d18b6ab)