Provides extension point to access to individual measurements
As suggested in #595.
MetricRegistry.addMeasurementListener() provides an extension point which lets users add listeners which will receive individual measurements for all registered metrics.
The registry holds a single MeasurementPublisher so that each metric object does not need to hold a separate collection.
Adds the metric name and a publisher to the signature of all metric constructors to allow them to emit measurements with the associated name.
Nested metrics do not report their measurements since these are meta-metrics—aggregations themselves.