Fix creating a uniform snapshot from a collection
The current code for creating a snapshot from the collection is unfortunately broken.
The code fails if the provided collection is modified during creating of the snapshot and
it has a weak iterator (precisely what `ConcurrentSkipListMap` provides). The iterator
doesn't guarantee that it has exactly the same amount iteration as the size of the
collection. As a result, the code throws an `ArrayIndexOutOfBoundException`.
The correct behaviour is too dump the collection to a resizable array on the fly and
then convert it the array of long primitives.

The behaviour was correct in the 3.1 branch, but was broken in 3.2 by the optimization
in #972.