Add ability to get default health check registry without an exception
Currently, there exists only one method for getting the default health
check registry - `getDefault`. Unfortunately, it throws an
`IllegalStateException` if the registry is not set. End users need to
check for this condition. Some users would like avoid using
exceptions for the control flow and would prefer to make a null check
instead catching an exception. For them we provide the `tryGetDefault`
method which does exactly that.