Adds a module for instrumentation of Jdbi 3 (#1229)
Jdbi3 has been released a couple weeks ago
(https://groups.google.com/forum/#!topic/jdbi/zS4p0p83sOM) and they use
a new model for instrumentation. This change adds support for setting
strategies for instrumenting Jdbi apps.

Implementation-wise this is a port of the `metrics-jdbi` module, improved,
upgraded to Java8, and with added support for the `@Timed` annotation
(previously it was not possible to set a custom metric name for an SqlObject
class and method).