Improve performance of ChunkedAssociativeLongArray
As suggested by errorprone, it's almost never a good idea to use a
`LinkedList`. The biggest problem with it is that it has a very big
cost of traversal (more 4 times bigger than the cost of removing an
elemt from an array). Considering iteration is more often than removing
in trimming, it makes sense just to use an array-based queue, because
it has a predictable-perfomance.

http://errorprone.info/bugpattern/JdkObsolete
https://stuartmarks.wordpress.com/2015/12/18/some-java-list-benchmarks/