Provide a type-safe way to create a custom counter
Practically speaking, the only way to customize a counter is too
extend it. That means, users who will provide a new supplier will
probably use the new class which extends `Counter`. It would be nice
if the `counter(MetricName name, MetricSupplier<Counter> supplier)`
method returned a counter with the same type as specified in the
`MetricSupplier`. This will save an redundant type-casting for the user
and avoid a potential compilation warning.