Add support for signing and verifying requests to CollectD
Starting with version 4.7 the CollectD provides the possibility to
cryptographically sign or encrypt the network traffic. The type
of protection is configured by the securityLevel option:

* None (plain traffic)
* Sign (packets should provide a digital signature)
* Encrypt (packets should be encrypted)

This changes adds possibility to set up a username, password,
and a securityLevel in the CollectD reporter, so it can send metrics
data  in the format expected by the CollectD server.

Resolves #1284