Jetty Update to 9.4.21
It contains some notable internal changes (most importantly related to
HttpChannel.State):
https://github.com/eclipse/jetty.project/commit/bde86467f4e5df595773ab11ed5e80c615b741f3#diff-68e74a60841dd5ec1f7edc4bd736ce18

Fix InstrumentedHandler to compile and work with pre-9.4.21 and
post-9.4.21 Jetty instances, but using internal API is not a recipe
for longevity.

Also, created a new InstrumentedHttpChannelListener that in turn
does use public API of Jetty, and should not break with minor
releases.