Add ability to send graphite metric attributes as tags.
Updates the GraphiteReporter to optionally send metric attributes as tags instead of using the default dot delimited format.
Defaults to the existing behavior, and does not attempt to add tag usage anywhere else within the project.

Primarily intended to allow other projects that use the graphite reporter as a library (micrometer) to support graphite tags.