Fix ExponentiallyDecayingReservoir constant rescaling based on the clock
This fixes a bug in ExponentiallyDecayingReservoir in which the
hour prior to the system precise clock (`System.nanotime()`) rolling
from `Long.MAX_VALUE` to `Long.MIN_VALUE` caused every interaction
with the reservoir to rescale.
It's easy to forget that the tick value is allowed to overflow.

In heavily loaded systems, this results in substantial performance
degredation due to write lock acquisition for each update.