Use Gauge<String> values as metric names
A lot of metric use gauges of type String to report the status of a component (e.g. failed / running). Currently all those values are discared since they are not numeric. This commit appends the value of the gauge to the metric name and sends it with a constant value of 1.