FEAT: Jersey: check impl method for annotations before interface method
Metrics annotations (@Timed, @Metered, etc) don’t allow annotating a
resource on the *implementing* method, only on the *defining* (interface)
method (or more accurately, the method corresponding to the @Path
definition, more or less).

This patch enhances dropwizard-metrics to extract metrics annotations more
intelligently, preferring the annotation on the implementation (and
falling back to the interface / definitionMethod if the implementation
annotation is absent).